Moved JetpackConnectionHelper to its own helper class
