+++ WordPress/src/main/java/org/wordpress/android/ui/jetpackrestconnection/JetpackConnector.kt
@@ -1,0 +1,9 @@
+package org.wordpress.android.ui.jetpackrestconnection
+
+import org.wordpress.android.fluxc.model.SiteModel
+
+class JetpackConnector {
+    suspend fun connectSite(site: SiteModel): Result<Boolean> {
+        return Result.success(true)
+    }
+}
